So many changes
Refactored solution structure (again, hopefully for good before going live).
Added/modified brainzzz's install script to work better for my needs. Now the script is a bit smarter in regards to not copying files if the solution is already in the rimworld mods folder and is more verbose about what it's doing in the output
Added first-pass compatibility with alien race framework. Tested with Xenohumans.